Phase 7 commit 28: Extract BookSignInRedirectHandler from MBDC
Lifts the SAML web-view + cookie-sync + post-auth retry path
(handleSAMLStartedState, handleLoginCancellation, handleBookFound,
handleProblem, clearAndSetCookies — ~180 LOC) into a focused
handler.

**Scope:**
- New: `Palace/MyBooks/BookSignInRedirectHandler.swift` (~265 LOC)
  with bookRegistry / stateManager / reauthenticator /
  userAccountProvider / shared credentialRequestState by injection.
  `BookSignInRedirectHandlerDelegate` exposes `cookieStorage`
  (computed via the active session) plus `cancelDownload(for:)` +
  `startDownload(for:withRequest:)`.
- `MyBooksDownloadCenter`: adds `signInRedirectHandler:
  BookSignInRedirectHandler` stored property + nil-default init
  param. Five inline methods deleted; `handleSAMLStartedState`,
  `handleProblem`, `clearAndSetCookies` become 1-line delegators
  (their callers in the borrow flow + `logInvalidURLRequest` keep
  working). New empty conformance + cookieStorage computed property
  on MBDC.
- pbxproj: 6 entries.

**Behavior parity (3 handleProblem branches preserved verbatim):**
- `.SAMLStarted` circuit breaker → sign-in modal without sign-out,
  shows problem-doc detail if present, gated by 2s
  isRequestingCredentials cooldown.
- SAML cookies expired (authDef.isSaml + hasCredentials) → cleanup
  + state .SAMLStarted + retry startDownload.
- No-credentials sign-in modal → reauthenticate + retry on
  hasCredentials.
- Same cancelDownload trigger on login-cancellation.
- Same userAccount.setCookies on bookFound.
- Same cookie-storage clear-then-set during clearAndSetCookies.

**DI improvement:**
The cookieStorage-via-delegate pattern is intentional. MBDC's
session is late-initialized after `super.init()`, so capturing
`self.session` in a closure during the handler's init triggers a
"self before super.init" error. Routing through the delegate's
computed cookieStorage property lets the late-bound session resolve
at call time AND makes test mocking tractable (a fake delegate can
return any HTTPCookieStorage instance).

**Net effect:**
- MBDC: 1,901 → 1,770 LOC (−131 LOC).
- 178-test smoke green (128 existing + 50 new service tests).

**Not done:**
- Direct BookSignInRedirectHandlerTests covering the 3 handleProblem
  branches with mocks.
- Mutation testing — palace_mutate.py worktree-path issue continues.

ForgeOS: cs_6a7e0e07 (init_e7603d4e) — review + testing + release
gates passed; can_release: true. Risk score 20 (auth-domain).
